+++ CHANGELOG.md
@@ -15,11 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - バイナリ内のレイヤーイメージを指すオブジェクトを使用する RasterImportedLayer が追加
 - レイヤーの追加に伴い、TTT PSD Importer が HSVAdjustmentLayer のインポート機能を追加 `#346`
 - TTT PSD Importer が SolidColorLayer のインポート機能を追加 `#346`
+- TextureSelector にモードが追加され、Absolute が追加 `#347`
+- TextureSelector にアバター内のテクスチャだけを列挙し、選択できる DomainTexturesSelector を追加 `#347`
 
 ### Changed
 
 - TTT PSD Importer はコンテキストメニューから、ScriptedImporter に変更 `#346`
 - SolidLayer は SolidColorLayer に名称変更 `#346`
+- TextureSelector にモードが追加され、以前までのデータは Relative に変更`#347`
+  - 上記に伴いフィールド名を変更 `#347`
 
 ### Removed